+++ src/cli/checkdb.go
@@ -29,7 +29,7 @@ func wrapDB(db *bolt.DB) *dbutil.DB {
 	return wdb
 }
 
-func checkdbCmd() *cobra.Command {
+func checkDBCmd() *cobra.Command {
 	return &cobra.Command{
 		Short: "Verify the database",
 		Use:   "checkdb [db path]",
@@ -38,23 +38,27 @@ func checkdbCmd() *cobra.Command {
 		Args:                  cobra.MaximumNArgs(1),
 		DisableFlagsInUseLine: true,
 		SilenceUsage:          true,
-		RunE:                  checkdb,
+		RunE:                  checkDB,
 	}
 }
 
-func checkdb(c *cobra.Command, args []string) error {
+func checkDB(c *cobra.Command, args []string) error {
 	// get db path
-	dbpath, err := resolveDBPath(cliConfig, args[0])
+	dbPath := ""
+	if len(args) > 0 {
+		dbPath = args[0]
+	}
+	dbPath, err := resolveDBPath(cliConfig, args[0])
 	if err != nil {
 		return err
 	}
 
-	// check if this file is exist
-	if _, err := os.Stat(dbpath); os.IsNotExist(err) {
-		return fmt.Errorf("db file: %v does not exist", dbpath)
+	// check if this file exists
+	if _, err := os.Stat(dbPath); os.IsNotExist(err) {
+		return fmt.Errorf("db file: %v does not exist", dbPath)
 	}
 
-	db, err := bolt.Open(dbpath, 0600, &bolt.Options{
+	db, err := bolt.Open(dbPath, 0600, &bolt.Options{
 		Timeout:  5 * time.Second,
 		ReadOnly: true,
 	})
@@ -80,4 +84,58 @@ func checkdb(c *cobra.Command, args []string) error {
 
 	fmt.Println("check db success")
 	return nil
+}
+
+func checkDBEncodingCmd() *cobra.Command {
+	return &cobra.Command{
+		Short: "Verify the database data encoding",
+		Use:   "checkDBDecoding [db path]",
+		Long: `Verify the generated binary encoders match the dynamic encoders for database data.
+    If no argument is specificed, the default data.db in $HOME/.$COIN/ will be checked.`,
+		Args:                  cobra.MaximumNArgs(1),
+		DisableFlagsInUseLine: true,
+		SilenceUsage:          true,
+		RunE:                  checkDBDecoding,
+	}
+}
+
+func checkDBDecoding(c *cobra.Command, args []string) error {
+	// get db path
+	dbPath := ""
+	if len(args) > 0 {
+		dbPath = args[0]
+	}
+	dbPath, err := resolveDBPath(cliConfig, dbPath)
+	if err != nil {
+		return err
+	}
+
+	// check if this file exists
+	if _, err := os.Stat(dbPath); os.IsNotExist(err) {
+		return fmt.Errorf("db file: %v does not exist", dbPath)
+	}
+
+	db, err := bolt.Open(dbPath, 0600, &bolt.Options{
+		Timeout:  5 * time.Second,
+		ReadOnly: true,
+	})
+
+	if err != nil {
+		return fmt.Errorf("open db failed: %v", err)
+	}
+
+	go func() {
+		apputil.CatchInterrupt(quitChan)
+	}()
+
+	if err := visor.VerifyDBSkyencoderSafe(wrapDB(db), quitChan); err != nil {
+		if err == visor.ErrVerifyStopped {
+			return nil
+		}
+		return fmt.Errorf("checkDBDecoding failed: %v", err)
+	}
+
+	fmt.Println("check db encoding success")
+	return nil
+
 }