+++ src/visor/historydb/output.go
@@ -9,6 +9,8 @@ import (
 	"github.com/skycoin/skycoin/src/visor/dbutil"
 )
 
+//go:generate skyencoder -struct UxOut
+
 // UxOutsBkt holds unspent outputs
 var UxOutsBkt = []byte("uxouts")
 
@@ -31,7 +33,9 @@ type ErrUxOutNotExist struct {
 
 // NewErrUxOutNotExist creates ErrUxOutNotExist from a UxID
 func NewErrUxOutNotExist(uxID string) error {
-	return ErrUxOutNotExist{UxID: uxID}
+	return ErrUxOutNotExist{
+		UxID: uxID,
+	}
 }
 
 func (e ErrUxOutNotExist) Error() string {
@@ -44,17 +48,31 @@ type uxOuts struct{}
 // put sets out value
 func (ux *uxOuts) put(tx *dbutil.Tx, out UxOut) error {
 	hash := out.Hash()
-	return dbutil.PutBucketValue(tx, UxOutsBkt, hash[:], encoder.Serialize(out))
+
+	n := EncodeSizeUxOut(&out)
+	buf := make([]byte, n)
+	if err := EncodeUxOut(buf, &out); err != nil {
+		return err
+	}
+
+	return dbutil.PutBucketValue(tx, UxOutsBkt, hash[:], buf)
 }
 
 // get gets UxOut of given id
 func (ux *uxOuts) get(tx *dbutil.Tx, uxID cipher.SHA256) (*UxOut, error) {
 	var out UxOut
 
-	if ok, err := dbutil.GetBucketObjectDecoded(tx, UxOutsBkt, uxID[:], &out); err != nil {
+	v, err := dbutil.GetBucketValueNoCopy(tx, UxOutsBkt, uxID[:])
+	if err != nil {
 		return nil, err
-	} else if !ok {
+	} else if v == nil {
 		return nil, nil
+	}
+
+	if n, err := DecodeUxOut(v, &out); err != nil {
+		return nil, err
+	} else if n != len(v) {
+		return nil, encoder.ErrRemainingBytes
 	}
 
 	return &out, nil